LayoutTests/fast/css/resize-single-axis-expected.txt

'both' resized as expected to (100px, 100px).
'horizontal' resized as expected to (100px, 50px).
'vertical' resized as expected to (50px, 100px).
'block-horizontal' resized as expected to (50px, 100px).
'inline-horizontal' resized as expected to (100px, 50px).
'block-vertical' resized as expected to (100px, 50px).
'inline-vertical' resized as expected to (50px, 100px).
